refactor provider form code
