+++ admin/client/config.ts
@@ -1,83 +1,16 @@
 import { z } from "zod";
+import {
+  AzureOpenAIConfigSchema,
+  DEFAULT_AZURE_OPENAI_CONFIG,
+} from "./providers/azure";
+import { DEFAULT_GEMINI_CONFIG, GeminiConfigSchema } from "./providers/gemini";
+import { DEFAULT_OLLAMA_CONFIG, OllamaConfigSchema } from "./providers/ollama";
+import { DEFAULT_OPENAI_CONFIG, OpenAIConfigSchema } from "./providers/openai";
 import { getBaseURL } from "./utils";
 
 // Rag config scheme
 export const RagConfigSchema = z.object({
   system_prompt: z.string().nullable().optional(),
-});
-
-// Model config schemes
-const BaseConfigSchema = z.object({
-  model_provider: z.string().nullable().optional(),
-  model: z.string().nullable().optional(),
-  embedding_model: z.string().nullable().optional(),
-  embedding_dim: z.number().nullable().optional(),
-  configured: z.boolean().nullable().optional(),
-});
-const OpenAIConfigSchema = BaseConfigSchema.extend({
-  model_provider: z.literal("openai").nullable().optional(),
-  openai_api_key: z
-    .string()
-    .nullable()
-    .optional()
-    .refine(
-      (value) => value && value.trim() !== "",
-      "OpenAI API Key is required",
-    ),
-});
-const GeminiConfigSchema = BaseConfigSchema.extend({
-  model_provider: z.literal("gemini").nullable().optional(),
-  google_api_key: z
-    .string()
-    .nullable()
-    .optional()
-    .refine(
-      (value) => value && value.trim() !== "",
-      "Google API Key is required",
-    ),
-});
-const OllamaConfigSchema = BaseConfigSchema.extend({
-  model_provider: z.literal("ollama").nullable().optional(),
-  ollama_base_url: z
-    .string()
-    .default("http://host.docker.internal:11434")
-    .nullable()
-    .optional(),
-});
-const AzureOpenAIConfigSchema = BaseConfigSchema.extend({
-  model_provider: z.literal("azure-openai").nullable().optional(),
-  azure_openai_endpoint: z
-    .string()
-    .nullable()
-    .optional()
-    .refine(
-      (value) => value && value.trim() !== "",
-      "Azure OpenAI endpoint is required",
-    ),
-  azure_openai_api_key: z
-    .string()
-    .nullable()
-    .optional()
-    .refine(
-      (value) => value && value.trim() !== "",
-      "Azure OpenAI API key is required",
-    ),
-  azure_openai_llm_deployment: z
-    .string()
-    .nullable()
-    .optional()
-    .refine(
-      (value) => value && value.trim() !== "",
-      "Azure OpenAI LLM deployment name is required",
-    ),
-  azure_openai_embedding_deployment: z
-    .string()
-    .nullable()
-    .optional()
-    .refine(
-      (value) => value && value.trim() !== "",
-      "Azure OpenAI embedding deployment name is required",
-    ),
 });
 
 // Merge the model config schemes with the Rag config scheme
@@ -123,36 +56,6 @@ export const supportedProviders = [
     value: "azure-openai",
   },
 ];
-
-export const DEFAULT_OPENAI_CONFIG: z.input<typeof OpenAIConfigSchema> = {
-  model_provider: "openai",
-  model: "gpt-3.5-turbo",
-  embedding_model: "text-embedding-3-small",
-  embedding_dim: 1536,
-  openai_api_key: "",
-};
-export const DEFAULT_GEMINI_CONFIG: z.input<typeof GeminiConfigSchema> = {
-  model_provider: "gemini",
-  model: "gemini-1.5-pro-latest",
-  embedding_model: "embedding-001",
-  embedding_dim: 768,
-  google_api_key: "",
-};
-export const DEFAULT_OLLAMA_CONFIG: z.input<typeof OllamaConfigSchema> = {
-  model_provider: "ollama",
-  model: "phi3:latest",
-  embedding_model: "nomic-embed-text",
-  embedding_dim: 768,
-  ollama_base_url: "http://host.docker.internal:11434",
-};
-export const DEFAULT_AZURE_OPENAI_CONFIG: z.input<
-  typeof AzureOpenAIConfigSchema
-> = {
-  model_provider: "azure-openai",
-  model: "gpt-35-turbo",
-  embedding_model: "text-embedding-3-small",
-  embedding_dim: 1536,
-};
 
 export const DEFAULT_CONFIG: z.input<typeof ConfigFormSchema> = {};
